<img src="https://github.com/smarras79/Jexpresso/blob/un/ci/assets/logo-ext2.png" width="500" title="JEXPRESSO logo">

Jexpresso

Jexpresso — Module

A research software for the numerical solution of a system of arbitrary conservation laws using continuous spectral elements.

Suggested Julia version: 1.10

If you are interested in contributing, please get in touch. Simone Marras, Yassine Tissaoui

Tables

TABLES